#106: Better handling how to determine new user, when sending welcome email.
